+++ packages/syntax-core/src/Card/Card.test.tsx
@@ -1,5 +1,6 @@
 import { screen, render } from "@testing-library/react";
 import Card from "./Card";
+import Box from "../Box/Box";
 
 describe("card", () => {
   it("should render children successfully", () => {
@@ -42,6 +43,23 @@ describe("card", () => {
     expect(screen.getByTestId("card-lg")).toHaveStyle({
       width: "100%",
       maxWidth: "744px",
+    });
+  });
+
+  it("should stretch to the width of the container", () => {
+    render(
+      <Box width={1000}>
+        <Card data-testid="card">
+          <>
+            <h1>title</h1>
+            <p>text</p>
+          </>
+        </Card>
+      </Box>,
+    );
+    expect(screen.getByTestId("card")).toHaveStyle({
+      width: "100%",
+      maxWidth: undefined,
     });
   });
 });