+++ packages/syntax-core/src/MiniActionCardGroup/MiniActionCardGroup.tsx
@@ -1,0 +1,11 @@
+import React, { type ReactNode } from "react";
+
+import styles from "./MiniActionCardGroup.module.css";
+
+export default function MiniActionCardGroup({
+  children,
+}: {
+  children: ReactNode;
+}): JSX.Element {
+  return <div className={styles.miniActionCardGroup}>{children}</div>;
+}